auth_service/requirements.txt
cryptography
uvicorn
gunicorn
fastapi
PyJWT
sqlalchemy
pymysql
python-dotenv